callback_learning_rate_scheduler {keras}	R Documentation

Learning rate scheduler.

Description

Learning rate scheduler.

Usage

callback_learning_rate_scheduler(schedule)

Arguments

schedule
a function that takes an epoch index as input (integer, indexed from 0) and current learning rate and returns a new learning rate as output (float).

See Also

Other callbacks: callback_csv_logger(), callback_early_stopping(), callback_lambda(), callback_model_checkpoint(), callback_progbar_logger(), callback_reduce_lr_on_plateau(), callback_remote_monitor(), callback_tensorboard(), callback_terminate_on_naan()

[Package keras version 2.6.1 Index]