Configure a Keras model for training

Configure a Keras model for training

compile(object, optimizer, loss, metrics = NULL, loss_weights = NULL,
  sample_weight_mode = NULL, ...)

Arguments

object
Model object to compile.

optimizer
Name of optimizer or optimizer object.

loss
Name of objective function or objective function. If the model has multiple outputs, you can use a different loss on each output by passing a dictionary or a list of objectives. The loss value that will be minimized by the model will then be the sum of all individual losses.

metrics
List of metrics to be evaluated by the model during training and testing. Typically you will use metrics='accuracy'. To specify different metrics for different outputs of a multi-output model, you could also pass a named list such as metrics=list(output_a = 'accuracy').

loss_weights
Optional list specifying scalar coefficients to weight the loss contributions of different model outputs. The loss value that will be minimized by the model will then be the weighted sum of all indvidual losses, weighted by the loss_weights coefficients.

sample_weight_mode
If you need to do timestep-wise sample weighting (2D weights), set this to "temporal". NULL defaults to sample-wise weights (1D). If the model has multiple outputs, you can use a different sample_weight_mode on each output by passing a list of modes.

...
Additional named arguments passed to tf$Session$run.

See also

Other model functions: evaluate_generator, evaluate, fit_generator, fit, get_config, get_layer, keras_model_sequential, keras_model, pop_layer, predict.tensorflow.keras.engine.training.Model, predict_generator, predict_on_batch, predict_proba, summary.tensorflow.keras.engine.training.Model, train_on_batch