UnitNorm weight constraint

Constrains the weights incident to each hidden unit to have unit norm.

constraint_unitnorm(axis = 0)

Arguments

axis
The axis along which to calculate weight norms. For instance, in a dense layer the weight matrix has shape input_dim, output_dim, set axis to 0 to constrain each weight vector of length input_dim,. In a convolution 2D layer with dim_ordering="tf", the weight tensor has shape rows, cols, input_depth, output_depth, set axis to c(0, 1, 2) to constrain the weights of each filter tensor of size rows, cols, input_depth.

See also

Other constraints: constraint_maxnorm, constraint_minmaxnorm, constraint_nonneg