count_params {keras}	R Documentation

Count the total number of scalars composing the weights.

Description

Count the total number of scalars composing the weights.

Usage

count_params(object)

Arguments

object
Layer or model object

Value

An integer count

See Also

Other layer methods: get_config(), get_input_at(), get_weights(), reset_states()

[Package keras version 2.6.1 Index]