create_layer {keras}	R Documentation

Create a Keras Layer

Description

Create a Keras Layer

Usage

create_layer(layer_class, object, args = list())

Arguments

layer_class
Python layer class or R6 class of type KerasLayer

object
Object to compose layer with. This is either a keras_model_sequential() to add the layer to, or another Layer which this layer will call.

args
List of arguments to layer constructor function

Value

A Keras layer

Note

The object parameter can be missing, in which case the layer is created without a connection to an existing graph.

[Package keras version 2.6.1 Index]