Generates batches of augmented/normalized data from images in a directory

Generates batches of augmented/normalized data from images in a directory

# S3 method for tensorflow.contrib.keras.python.keras.preprocessing.image.ImageDataGenerator
flow_data_from_directory(generator,
  directory, target_size = c(256, 256), color_mode = "rgb",
  classes = NULL, class_mode = "categorical", batch_size = 32,
  shuffle = TRUE, seed = NULL, save_to_dir = NULL, save_prefix = "",
  save_format = "jpeg", follow_links = FALSE)

Arguments

generator
Image data generator

directory
path to the target directory. It should contain one subdirectory per class. Any PNG, JPG or BMP images inside each of the subdirectories directory tree will be included in the generator. See thisscript for more details.

target_size
integer vectir, default: c(256, 256). The dimensions to which all images found will be resized.

color_mode
one of "grayscale", "rbg". Default: "rgb". Whether the images will be converted to have 1 or 3 color channels.

classes
optional list of class subdirectories (e.g. c('dogs', 'cats')). Default: NULL, If not provided, the list of classes will be automatically inferred (and the order of the classes, which will map to the label indices, will be alphanumeric).

class_mode
one of "categorical", "binary", "sparse" or NULL. Default: "categorical". Determines the type of label arrays that are returned: "categorical" will be 2D one-hot encoded labels, "binary" will be 1D binary labels, "sparse" will be 1D integer labels. If NULL, no labels are returned (the generator will only yield batches of image data, which is useful to use predict_generator(), evaluate_generator(), etc.).

batch_size
int (default: 32).

shuffle
boolean (defaut: TRUE).

seed
int (default: NULL).

save_to_dir
NULL or str (default: NULL). This allows you to optimally specify a directory to which to save the augmented pictures being generated (useful for visualizing what you are doing).

save_prefix
str (default: ''). Prefix to use for filenames of saved pictures (only relevant if save_to_dir is set).

save_format
one of "png", "jpeg" (only relevant if save_to_dir is set). Default: "jpeg".

follow_links
whether to follow symlinks inside class subdirectories (default: FALSE).

...
Unused

Details

Yields batches indefinitely, in an infinite loop.

Yields

(x, y) where x is an array of image data and y is a array of corresponding labels. The generator loops indefinitely.

See also

Other image preprocessing: fit.tensorflow.contrib.keras.python.keras.preprocessing.image.ImageDataGenerator, flow_data.tensorflow.contrib.keras.python.keras.preprocessing.image.ImageDataGenerator, image_load, image_to_array