image_load {keras}	R Documentation

Loads an image into PIL format.

Description

Loads an image into PIL format.

Usage

image_load(
  path,
  grayscale = FALSE,
  target_size = NULL,
  interpolation = "nearest"
)

Arguments

path
Path to image file

grayscale
Boolean, whether to load the image as grayscale.

target_size
Either NULL (default to original size) or integer vector (img_height, img_width).

interpolation
Interpolation method used to resample the image if the target size is different from that of the loaded image. Supported methods are "nearest", "bilinear", and "bicubic". If PIL version 1.1.3 or newer is installed, "lanczos" is also supported. If PIL version 3.4.0 or newer is installed, "box" and "hamming" are also supported. By default, "nearest" is used.

Value

A PIL Image instance.

See Also

Other image preprocessing: fit_image_data_generator(), flow_images_from_dataframe(), flow_images_from_data(), flow_images_from_directory(), image_to_array()

[Package keras version 2.6.1 Index]