k_arange {keras}	R Documentation

Creates a 1D tensor containing a sequence of integers.

Description

The function arguments use the same convention as Theano's arange: if only one argument is provided, it is in fact the "stop" argument. The default type of the returned tensor is 'int32' to match TensorFlow's default.

Usage

k_arange(start, stop = NULL, step = 1, dtype = "int32")

Arguments

start
Start value.

stop
Stop value.

step
Difference between two successive values.

dtype
Integer dtype to use.

Value

An integer tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]