Returns the index of the maximum value along an axis.

Returns the index of the maximum value along an axis.

k_argmax(x, axis = -1L)

Arguments

x
Tensor or variable.

axis
axis along which to perform the reduction.

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.