k_batch_get_value {keras}	R Documentation

Returns the value of more than one tensor variable.

Description

Returns the value of more than one tensor variable.

Usage

k_batch_get_value(ops)

Arguments

ops
List of ops to evaluate.

Value

A list of arrays.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

See Also

k_batch_set_value()

[Package keras version 2.6.1 Index]