k_batch_normalization {keras}	R Documentation

Applies batch normalization on x given mean, var, beta and gamma.

Description

i.e. returns output <- (x - mean) / (sqrt(var) + epsilon) * gamma + beta

Usage

k_batch_normalization(x, mean, var, beta, gamma, axis = -1, epsilon = 0.001)

Arguments

x
Input tensor or variable.

mean
Mean of batch.

var
Variance of batch.

beta
Tensor with which to center the input.

gamma
Tensor by which to scale the input.

axis
Axis (axis indexes are 1-based). Pass -1 (the default) to select the last axis.

epsilon
Fuzz factor.

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]