k_binary_crossentropy {keras}	R Documentation

Binary crossentropy between an output tensor and a target tensor.

Description

Binary crossentropy between an output tensor and a target tensor.

Usage

k_binary_crossentropy(target, output, from_logits = FALSE)

Arguments

target
A tensor with the same shape as output.

output
A tensor.

from_logits
Whether output is expected to be a logits tensor. By default, we consider that output encodes a probability distribution.

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]