k_cast {keras}	R Documentation

Casts a tensor to a different dtype and returns it.

Description

You can cast a Keras variable but it still returns a Keras tensor.

Usage

k_cast(x, dtype)

Arguments

x
Keras tensor (or variable).

dtype
String, either ('float16', 'float32', or 'float64').

Value

Keras tensor with dtype dtype.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]