k_categorical_crossentropy {keras}	R Documentation

Categorical crossentropy between an output tensor and a target tensor.

Description

Categorical crossentropy between an output tensor and a target tensor.

Usage

k_categorical_crossentropy(target, output, from_logits = FALSE, axis = -1)

Arguments

target
A tensor of the same shape as output.

output
A tensor resulting from a softmax (unless from_logits is TRUE, in which case output is expected to be the logits).

from_logits
Logical, whether output is the result of a softmax, or is a tensor of logits.

axis
Axis (axis indexes are 1-based). Pass -1 (the default) to select the last axis.

Value

Output tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]