k_conv2d_transpose {keras}	R Documentation

2D deconvolution (i.e. transposed convolution).

Description

2D deconvolution (i.e. transposed convolution).

Usage

k_conv2d_transpose(
  x,
  kernel,
  output_shape,
  strides = c(1, 1),
  padding = "valid",
  data_format = NULL
)

Arguments

x
Tensor or variable.

kernel
kernel tensor.

output_shape
1D int tensor for the output shape.

strides
strides list.

padding
string, "same" or "valid".

data_format
string, "channels_last" or "channels_first". Whether to use Theano or TensorFlow/CNTK data format for inputs/kernels/outputs.

Value

A tensor, result of transposed 2D convolution.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]