k_ctc_batch_cost {keras}	R Documentation

Runs CTC loss algorithm on each batch element.

Description

Runs CTC loss algorithm on each batch element.

Usage

k_ctc_batch_cost(y_true, y_pred, input_length, label_length)

Arguments

y_true
tensor (samples, max_string_length) containing the truth labels.

y_pred
tensor (samples, time_steps, num_categories) containing the prediction, or output of the softmax.

input_length
tensor (samples, 1) containing the sequence length for each batch item in y_pred.

label_length
tensor (samples, 1) containing the sequence length for each batch item in y_true.

Value

Tensor with shape (samples,1) containing the CTC loss of each element.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]