k_ctc_decode {keras}	R Documentation

Decodes the output of a softmax.

Description

Can use either greedy search (also known as best path) or a constrained dictionary search.

Usage

k_ctc_decode(
  y_pred,
  input_length,
  greedy = TRUE,
  beam_width = 100L,
  top_paths = 1
)

Arguments

y_pred
tensor (samples, time_steps, num_categories) containing the prediction, or output of the softmax.

input_length
tensor (samples, ) containing the sequence length for each batch item in y_pred.

greedy
perform much faster best-path search if TRUE. This does not use a dictionary.

beam_width
if greedy is FALSE: a beam search decoder will be used with a beam of this width.

top_paths
if greedy is FALSE, how many of the most probable paths will be returned.

Value

If greedy is TRUE, returns a list of one element that contains the decoded sequence. If FALSE, returns the top_paths most probable decoded sequences. Important: blank labels are returned as -1. Tensor (top_paths) that contains the log probability of each decoded sequence.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]