k_ctc_label_dense_to_sparse {keras}	R Documentation

Converts CTC labels from dense to sparse.

Description

Converts CTC labels from dense to sparse.

Usage

k_ctc_label_dense_to_sparse(labels, label_lengths)

Arguments

labels
dense CTC labels.

label_lengths
length of the labels.

Value

A sparse tensor representation of the labels.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]