k_dropout {keras}	R Documentation

Sets entries in x to zero at random, while scaling the entire tensor.

Description

Sets entries in x to zero at random, while scaling the entire tensor.

Usage

k_dropout(x, level, noise_shape = NULL, seed = NULL)

Arguments

x
tensor

level
fraction of the entries in the tensor that will be set to 0.

noise_shape
shape for randomly generated keep/drop flags, must be broadcastable to the shape of x

seed
random seed to ensure determinism.

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]