k_epsilon {keras}	R Documentation

Fuzz factor used in numeric expressions.

Description

Fuzz factor used in numeric expressions.

Usage

k_epsilon()

k_set_epsilon(e)

Arguments

e
float. New value of epsilon.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]