k_exp {keras}	R Documentation

Element-wise exponential.

Description

Element-wise exponential.

Usage

k_exp(x)

Arguments

x
Tensor or variable.

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]