k_expand_dims {keras}	R Documentation

Adds a 1-sized dimension at index axis.

Description

Adds a 1-sized dimension at index axis.

Usage

k_expand_dims(x, axis = -1)

Arguments

x
A tensor or variable.

axis
Position where to add a new axis (axis indexes are 1-based). Pass -1 (the default) to select the last axis.

Value

A tensor with expanded dimensions.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]