k_foldr {keras}	R Documentation

Reduce elems using fn to combine them from right to left.

Description

Reduce elems using fn to combine them from right to left.

Usage

k_foldr(fn, elems, initializer = NULL, name = NULL)

Arguments

fn
Function that will be called upon each element in elems and an accumulator

elems
tensor

initializer
The first value used (last element of elems in case of NULL)

name
A string name for the foldr node in the graph

Value

Tensor with same type and shape as initializer.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]