k_function {keras}	R Documentation

Instantiates a Keras function

Description

Instantiates a Keras function

Usage

k_function(inputs, outputs, updates = NULL, ...)

Arguments

inputs
List of placeholder tensors.

outputs
List of output tensors.

updates
List of update ops.

...
Named arguments passed to tf$Session$run.

Value

Output values as R arrays.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]