k_gather {keras}	R Documentation

Retrieves the elements of indices indices in the tensor reference.

Description

Retrieves the elements of indices indices in the tensor reference.

Usage

k_gather(reference, indices)

Arguments

reference
A tensor.

indices
Indices. Dimension indices are 1-based. Note however that if you pass a tensor for indices they will be passed as-is, in which case indices will be 0 based because no normalizing of R 1-based axes to Python 0-based axes is performed.

Value

A tensor of same type as reference.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]