k_image_data_format {keras}	R Documentation

Default image data format convention ('channels_first' or 'channels_last').

Description

Default image data format convention ('channels_first' or 'channels_last').

Usage

k_image_data_format()

k_set_image_data_format(data_format)

Arguments

data_format
string. 'channels_first' or 'channels_last'.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]