k_in_test_phase {keras}	R Documentation

Selects x in test phase, and alt otherwise.

Description

Note that alt should have the same shape as x.

Usage

k_in_test_phase(x, alt, training = NULL)

Arguments

x
What to return in test phase (tensor or function that returns a tensor).

alt
What to return otherwise (tensor or function that returns a tensor).

training
Optional scalar tensor (or R logical or integer) specifying the learning phase.

Value

Either x or alt based on k_learning_phase().

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]