k_is_sparse {keras}	R Documentation

Returns whether a tensor is a sparse tensor.

Description

Returns whether a tensor is a sparse tensor.

Usage

k_is_sparse(tensor)

Arguments

tensor
A tensor instance.

Value

A logical

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]