k_less_equal {keras}	R Documentation

Element-wise truth value of (x <= y).

Description

Element-wise truth value of (x <= y).

Usage

k_less_equal(x, y)

Arguments

x
Tensor or variable.

y
Tensor or variable.

Value

A bool tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]