k_logsumexp {keras}	R Documentation

Computes log(sum(exp(elements across dimensions of a tensor))).

Description

This function is more numerically stable than log(sum(exp(x))). It avoids overflows caused by taking the exp of large inputs and underflows caused by taking the log of small inputs.

Usage

k_logsumexp(x, axis = NULL, keepdims = FALSE)

Arguments

x
A tensor or variable.

axis
An integer, the axis to reduce over (axis indexes are 1-based).

keepdims
A boolean, whether to keep the dimensions or not. If keepdims is FALSE, the rank of the tensor is reduced by 1. If keepdims is TRUE, the reduced dimension is retained with length 1.

Value

The reduced tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]