k_manual_variable_initialization {keras}	R Documentation

Sets the manual variable initialization flag.

Description

This boolean flag determines whether variables should be initialized as they are instantiated (default), or if the user should handle the initialization (e.g. via tf$initialize_all_variables()).

Usage

k_manual_variable_initialization(value)

Arguments

value
Logical

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]