k_normalize_batch_in_training {keras}	R Documentation

Computes mean and std for batch then apply batch_normalization on batch.

Description

Computes mean and std for batch then apply batch_normalization on batch.

Usage

k_normalize_batch_in_training(x, gamma, beta, reduction_axes, epsilon = 0.001)

Arguments

x
Input tensor or variable.

gamma
Tensor by which to scale the input.

beta
Tensor with which to center the input.

reduction_axes
iterable of integers, axes over which to normalize.

epsilon
Fuzz factor.

Value

A list length of 3, (normalized_tensor, mean, variance).

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]