k_ones_like {keras}	R Documentation

Instantiates an all-ones variable of the same shape as another tensor.

Description

Instantiates an all-ones variable of the same shape as another tensor.

Usage

k_ones_like(x, dtype = NULL, name = NULL)

Arguments

x
Keras variable or tensor.

dtype
String, dtype of returned Keras variable. NULL uses the dtype of x.

name
String, name for the variable to create.

Value

A Keras variable with the shape of x filled with ones.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]