k_placeholder {keras}	R Documentation

Instantiates a placeholder tensor and returns it.

Description

Instantiates a placeholder tensor and returns it.

Usage

k_placeholder(
  shape = NULL,
  ndim = NULL,
  dtype = NULL,
  sparse = FALSE,
  name = NULL
)

Arguments

shape
Shape of the placeholder (integer list, may include NULL entries).

ndim
Number of axes of the tensor. At least one of shape, ndim must be specified. If both are specified, shape is used.

dtype
Placeholder type.

sparse
Logical, whether the placeholder should have a sparse type.

name
Optional name string for the placeholder.

Value

Tensor instance (with Keras metadata included).

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]