k_prod {keras}	R Documentation

Multiplies the values in a tensor, alongside the specified axis.

Description

Multiplies the values in a tensor, alongside the specified axis.

Usage

k_prod(x, axis = NULL, keepdims = FALSE)

Arguments

x
A tensor or variable.

axis
An integer, axis to compute the product over (axis indexes are 1-based).

keepdims
A boolean, whether to keep the dimensions or not. If keepdims is FALSE, the rank of the tensor is reduced by 1. If keepdims is TRUE, the reduced dimension is retained with length 1.

Value

A tensor with the product of elements of x.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]