k_repeat_elements {keras}	R Documentation

Repeats the elements of a tensor along an axis.

Description

If x has shape (s1, s2, s3) and axis is 2, the output will have shape (s1, s2 * rep, s3).

Usage

k_repeat_elements(x, rep, axis)

Arguments

x
Tensor or variable.

rep
Integer, number of times to repeat.

axis
Axis along which to repeat (axis indexes are 1-based)

Value

A tensor.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]