k_squeeze {keras}	R Documentation

Removes a 1-dimension from the tensor at index axis.

Description

Removes a 1-dimension from the tensor at index axis.

Usage

k_squeeze(x, axis)

Arguments

x
A tensor or variable.

axis
Axis to drop (axis indexes are 1-based).

Value

A tensor with the same data as x but reduced dimensions.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]