k_update_sub {keras}	R Documentation

Update the value of x by subtracting decrement.

Description

Update the value of x by subtracting decrement.

Usage

k_update_sub(x, decrement)

Arguments

x
A Variable.

decrement
A tensor of same shape as x.

Value

The variable x updated.

Keras Backend

This function is part of a set of Keras backend functions that enable lower level access to the core operations of the backend tensor engine (e.g. TensorFlow, CNTK, Theano, etc.).

You can see a list of all available backend functions here: https://keras.rstudio.com/articles/backend.html#backend-functions.

[Package keras version 2.6.1 Index]