R interface to Keras

Keras is a high-level neural networks API, developed with a focus on enabling fast experimentation. Keras has the following key features:

Details

Allows the same code to run on CPU or on GPU, seamlessly.

User-friendly API which makes it easy to quickly prototype deep learning models.

Built-in support for convolutional networks (for computer vision), recurrent networks (for sequence processing), and any combination of both.

Supports arbitrary network architectures: multi-input or multi-output models, layer sharing, model sharing, etc. This means that Keras is appropriate for building essentially any deep learning model, from a memory network to a neural Turing machine.

Is capable of running on top of multiple back-ends including TensorFlow, Jax, or PyTorch.

See the package website at https://keras.rstudio.com for complete documentation.

See also

Author

Maintainer: Tomasz Kalinowski tomasz@posit.co [contributor, copyright holder]

Authors:

JJ Allaire [copyright holder]

François Chollet [copyright holder]

Other contributors:

Daniel Falbel daniel@rstudio.com [contributor, copyright holder]

RStudio [contributor, copyright holder, funder]

Google [contributor, copyright holder, funder]

Yuan Tang terrytangyuan@gmail.com (ORCID) [contributor, copyright holder]

Wouter Van Der Bijl [contributor, copyright holder]

Martin Studer [contributor, copyright holder]

Sigrid Keydana [contributor]