layer_average {keras}	R Documentation

Layer that averages a list of inputs.

Description

It takes as input a list of tensors, all of the same shape, and returns a single tensor (also of the same shape).

Usage

layer_average(inputs, ...)

Arguments

inputs
A list of input tensors (at least 2). Can be missing.

...
Standard layer arguments (must be named).

Value

A tensor, the average of the inputs. If inputs is missing, a keras layer instance is returned.

See Also

https://www.tensorflow.org/api_docs/python/tf/keras/layers/average

https://www.tensorflow.org/api_docs/python/tf/keras/layers/Average

https://keras.io/api/layers/merging_layers/average

Other merge layers: layer_concatenate(), layer_dot(), layer_maximum(), layer_minimum(), layer_multiply(), layer_subtract()

[Package keras version 2.6.1 Index]