layer_center_crop {keras}	R Documentation

Crop the central portion of the images to target height and width

Description

Crop the central portion of the images to target height and width

Usage

layer_center_crop(object, height, width, ...)

Arguments

object
What to call the new Layer instance with. Typically a keras Model, another Layer, or a tf.Tensor/KerasTensor. If object is missing, the Layer instance is returned, otherwise, layer(object) is returned.

height
Integer, the height of the output shape.

width
Integer, the width of the output shape.

...
standard layer arguments.

Details

Input shape: 3D (unbatched) or 4D (batched) tensor with shape: (..., height, width, channels), in "channels_last" format.

Output shape: 3D (unbatched) or 4D (batched) tensor with shape: (..., target_height, target_width, channels).

If the input height/width is even and the target height/width is odd (or inversely), the input image is left-padded by 1 pixel.

See Also

https://www.tensorflow.org/api_docs/python/tf/keras/layers/CenterCrop

https://keras.io/api/layers/preprocessing_layers/image_preprocessing/center_crop

Other image preprocessing layers: layer_rescaling(), layer_resizing()

Other preprocessing layers: layer_category_encoding(), layer_discretization(), layer_hashing(), layer_integer_lookup(), layer_normalization(), layer_random_contrast(), layer_random_crop(), layer_random_flip(), layer_random_height(), layer_random_rotation(), layer_random_translation(), layer_random_width(), layer_random_zoom(), layer_rescaling(), layer_resizing(), layer_string_lookup(), layer_text_vectorization()

[Package keras version 2.6.1 Index]