Cropping layer for 2D input (e.g. picture).

It crops along spatial dimensions, i.e. width and height.

layer_cropping_2d(x, cropping = list(c(0L, 0L), c(0L, 0L)),
  data_format = NULL)

Arguments

x
Model or layer

cropping
int, or list of 2 ints, or list of 2 lists of 2 ints.

If int: the same symmetric cropping is applied to width and height.

If list of 2 ints: interpreted as two different symmetric cropping values for height and width: (symmetric_height_crop, symmetrc_width_crop).

If list of 2 lists of 2 ints: interpreted as ((top_crop, bottom_crop), (left_crop, right_crop))

data_format
A string, one of channels_last (default) or channels_first. The ordering of the dimensions in the inputs. channels_last corresponds to inputs with shape (batch, width, height, channels) while channels_first corresponds to inputs with shape (batch, channels, width, height). It defaults to the image_data_format value found in your Keras config file at ~/.keras/keras.json. If you never set it, then it will be "channels_last".

Input shape

4D tensor with shape:

If data_format is "channels_last": (batch, rows, cols, channels)

If data_format is "channels_first": (batch, channels, rows, cols)

Output shape

4D tensor with shape:

If data_format is "channels_last": (batch, cropped_rows, cropped_cols, channels)

If data_format is "channels_first": (batch, channels, cropped_rows, cropped_cols)

See also

Other convolutional layers: layer_conv_1d, layer_conv_2d_transpose, layer_conv_2d, layer_conv_3d, layer_conv_lstm_2d, layer_cropping_1d, layer_cropping_3d, layer_separable_conv_2d, layer_upsampling_1d, layer_upsampling_2d, layer_upsampling_3d, layer_zero_padding_1d, layer_zero_padding_2d, layer_zero_padding_3d