Layer that computes a dot product between samples in two tensors.

Layer that computes a dot product between samples in two tensors.

layer_dot(inputs, axes, normalize = FALSE)

Arguments

inputs
A list of input tensors (at least 2).

axes
Integer or list of integers, axis or axes along which to take the dot product.

normalize
Whether to L2-normalize samples along the dot product axis before taking the dot product. If set to TRUE, then the output of the dot product is the cosine proximity between the two samples. **kwargs: Standard layer keyword arguments.

Value

A tensor, the dot product of the samples from the inputs.

See also

Other merge layers: layer_add, layer_average, layer_concatenate, layer_maximum, layer_multiply, layer_subtract