Turns positive integers (indexes) into dense vectors of fixed size.

e.g. rbind(4L, 20L) \(\rightarrow\) rbind(c(0.25, 0.1), c(0.6, -0.2))

This layer can only be used on positive integer inputs of a fixed range.

layer_embedding(
  object,
  input_dim,
  output_dim,
  embeddings_initializer = "uniform",
  embeddings_regularizer = NULL,
  embeddings_constraint = NULL,
  mask_zero = FALSE,
  ...
)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

input_dim

Integer. Size of the vocabulary, i.e. maximum integer index + 1.

output_dim

Integer. Dimension of the dense embedding.

embeddings_initializer

Initializer for the embeddings matrix (see keras::initializer_*).

embeddings_regularizer

Regularizer function applied to the embeddings matrix (see keras::regularizer_*).

embeddings_constraint

Constraint function applied to the embeddings matrix (see keras::constraint_*).

mask_zero

Boolean, whether or not the input value 0 is a special "padding" value that should be masked out. This is useful when using recurrent layers which may take variable length input. If this is TRUE, then all subsequent layers in the model need to support masking or an exception will be raised. If mask_zero is set to TRUE, as a consequence, index 0 cannot be used in the vocabulary (input_dim should equal size of vocabulary + 1).

...

For forward/backward compatability.

Examples

model <- keras_model_sequential() |>
  layer_embedding(1000, 64)

# The model will take as input an integer matrix of size (batch,input_length),
# and the largest integer (i.e. word index) in the input
# should be no larger than 999 (vocabulary size).
# Now model$output_shape is (NA, 10, 64), where `NA` is the batch
# dimension.

input_array <- random_integer(shape = c(32, 10), minval = 0, maxval = 1000)
model |> compile('rmsprop', 'mse')
output_array <- model |> predict(input_array, verbose = 0)
dim(output_array)    # (32, 10, 64)

## [1] 32 10 64

Input Shape

2D tensor with shape: (batch_size, input_length).

Output Shape

3D tensor with shape: (batch_size, input_length, output_dim).

See also