Flattens the input. Does not affect the batch size.

Flattens the input. Does not affect the batch size.

layer_flatten(object, data_format = NULL, ...)

Arguments

object

Object to compose the layer with. A tensor, array, or sequential model.

data_format

A string, one of "channels_last" (default) or "channels_first". The ordering of the dimensions in the inputs. "channels_last" corresponds to inputs with shape (batch, ..., channels) while "channels_first" corresponds to inputs with shape (batch, channels, ...). When unspecified, uses image_data_format value found in your Keras config file at ~/.keras/keras.json (if exists). Defaults to "channels_last".

...

Passed on to the Python callable

Note

If inputs are shaped (batch) without a feature axis, then flattening adds an extra channel dimension and output shape is (batch, 1).

Examples

x <- layer_input(shape=c(10, 64))
y <- x |> layer_flatten()
y$shape

## [[1]]
## NULL
##
## [[2]]
## [1] 640

See also