layer_random_crop {keras}	R Documentation

Randomly crop the images to target height and width

Description

Randomly crop the images to target height and width

Usage

layer_random_crop(object, height, width, seed = NULL, ...)

Arguments

object
What to call the new Layer instance with. Typically a keras Model, another Layer, or a tf.Tensor/KerasTensor. If object is missing, the Layer instance is returned, otherwise, layer(object) is returned.

height
Integer, the height of the output shape.

width
Integer, the width of the output shape.

seed
Integer. Used to create a random seed.

...
standard layer arguments.

Details

This layer will crop all the images in the same batch to the same cropping location. By default, random cropping is only applied during training. At inference time, the images will be first rescaled to preserve the shorter side, and center cropped. If you need to apply random cropping at inference time, set training to TRUE when calling the layer.

Input shape: 3D (unbatched) or 4D (batched) tensor with shape: (..., height, width, channels), in "channels_last" format.

Output shape: 3D (unbatched) or 4D (batched) tensor with shape: (..., target_height, target_width, channels).

See Also

https://www.tensorflow.org/api_docs/python/tf/keras/layers/RandomCrop

https://keras.io/api/layers/preprocessing_layers/image_augmentation/random_crop

Other image augmentation layers: layer_random_contrast(), layer_random_flip(), layer_random_height(), layer_random_rotation(), layer_random_translation(), layer_random_width(), layer_random_zoom()

Other preprocessing layers: layer_category_encoding(), layer_center_crop(), layer_discretization(), layer_hashing(), layer_integer_lookup(), layer_normalization(), layer_random_contrast(), layer_random_flip(), layer_random_height(), layer_random_rotation(), layer_random_translation(), layer_random_width(), layer_random_zoom(), layer_rescaling(), layer_resizing(), layer_string_lookup(), layer_text_vectorization()

[Package keras version 2.6.1 Index]