metric_cosine_proximity {keras}	R Documentation

(Deprecated) metric_cosine_proximity

Description

metric_cosine_proximity() is deprecated and will be removed in a future version. Please update your code to use metric_cosine_similarity() if possible. If you need the actual function and not a Metric object, (e.g, because you are using the intermediate computed values in a custom training loop before reduction), please use loss_cosine_similarity() or tensorflow::tf$compat$v1$keras$metrics$cosine_proximity()

Usage

metric_cosine_proximity(y_true, y_pred)

Arguments

y_true
Tensor of true targets.

y_pred
Tensor of predicted targets.

[Package keras version 2.6.1 Index]