model_from_saved_model {keras}	R Documentation

Load a Keras model from the Saved Model format

Description

Load a Keras model from the Saved Model format

Usage

model_from_saved_model(saved_model_path, custom_objects = NULL)

Arguments

saved_model_path
a string specifying the path to the SavedModel directory.

custom_objects
Optional dictionary mapping string names to custom classes or functions (e.g. custom loss functions).

Value

a Keras model.

Note

This functionality is experimental and only works with TensorFlow version >= "2.0".

See Also

Other saved_model: model_to_saved_model()

[Package keras version 2.6.1 Index]