model_to_yaml {keras}	R Documentation

Model configuration as YAML

Description

Save and re-load models configurations as YAML Note that the representation does not include the weights, only the architecture.

Usage

model_to_yaml(object)

model_from_yaml(yaml, custom_objects = NULL)

Arguments

object
Model object to save

yaml
YAML with model configuration

custom_objects
Optional named list mapping names to custom classes or functions to be considered during deserialization.

See Also

Other model persistence: get_weights(), model_to_json(), save_model_hdf5(), save_model_tf(), save_model_weights_hdf5(), serialize_model()

[Package keras version 2.6.1 Index]