predict.keras.engine.training.Model {keras}	R Documentation

Generate predictions from a Keras model

Description

Generates output predictions for the input samples, processing the samples in a batched way.

Usage

## S3 method for class 'keras.engine.training.Model'
predict(
  object,
  x,
  batch_size = NULL,
  verbose = 0,
  steps = NULL,
  callbacks = NULL,
  ...
)

Arguments

object
Keras model

x
Input data (vector, matrix, or array). You can also pass a tfdataset or a generator returning a list with (inputs, targets) or (inputs, targets, sample_weights).

batch_size
Integer. If unspecified, it will default to 32.

verbose
Verbosity mode, 0 or 1.

steps
Total number of steps (batches of samples) before declaring the evaluation round finished. Ignored with the default value of NULL.

callbacks
List of callbacks to apply during prediction.

...
Unused

Value

vector, matrix, or array of predictions

See Also

Other model functions: compile.keras.engine.training.Model(), evaluate.keras.engine.training.Model(), evaluate_generator(), fit.keras.engine.training.Model(), fit_generator(), get_config(), get_layer(), keras_model_sequential(), keras_model(), multi_gpu_model(), pop_layer(), predict_generator(), predict_on_batch(), predict_proba(), summary.keras.engine.training.Model(), train_on_batch()

[Package keras version 2.6.1 Index]