timeseries_generator {keras}	R Documentation

Utility function for generating batches of temporal data.

Description

Utility function for generating batches of temporal data.

Usage

timeseries_generator(
  data,
  targets,
  length,
  sampling_rate = 1,
  stride = 1,
  start_index = 0,
  end_index = NULL,
  shuffle = FALSE,
  reverse = FALSE,
  batch_size = 128
)

Arguments

data
Object containing consecutive data points (timesteps). The data should be 2D, and axis 1 is expected to be the time dimension.

targets
Targets corresponding to timesteps in data. It should have same length as data.

length
Length of the output sequences (in number of timesteps).

sampling_rate
Period between successive individual timesteps within sequences. For rate r, timesteps data[i], data[i-r], ... data[i - length] are used for create a sample sequence.

stride
Period between successive output sequences. For stride s, consecutive output samples would be centered around data[i], data[i+s], data[i+2*s], etc.

start_index, end_index
Data points earlier than start_index or later than end_index will not be used in the output sequences. This is useful to reserve part of the data for test or validation.

shuffle
Whether to shuffle output samples, or instead draw them in chronological order.

reverse
Boolean: if true, timesteps in each output sample will be in reverse chronological order.

batch_size
Number of timeseries samples in each batch (except maybe the last one).

Value

An object that can be passed to generator based training functions (e.g. fit_generator()).ma

[Package keras version 2.6.1 Index]