KerasWrapper {keras}	R Documentation

(Deprecated) Base R6 class for Keras wrappers

Description

Instead of inheriting from the proxy class KerasWrapper and using create_wrapper to create instances, new R6 custom classes are encouraged to inherit directly from keras$layers$Wrapper and use create_layer to create instances.

Format

An R6Class generator object

Value

KerasWrapper.

Methods

build(input_shape)

Builds the wrapped layer. Subclasses can extend this to perform custom operations on that layer.

call(inputs,mask)

Calls the wrapped layer on an input tensor.

compute_output_shape(input_shape)

Computes the output shape for the wrapped layer.

add_loss(losses, inputs)

Subclasses can use this to add losses to the wrapped layer.

add_weight(name,shape,dtype,initializer,regularizer,trainable,constraint)

Subclasses can use this to add weights to the wrapped layer.

[Package keras version 2.6.1 Index]