Xception V1 model for Keras.

Xception V1 model for Keras.

application_xception(
  include_top = TRUE,
  weights = "imagenet",
  input_tensor = NULL,
  input_shape = NULL,
  pooling = NULL,
  classes = 1000
)

xception_preprocess_input(x)

Arguments

include_top
whether to include the fully-connected layer at the top of the network.

weights
NULL (random initialization), imagenet (ImageNet weights), or the path to the weights file to be loaded.

input_tensor
optional Keras tensor to use as image input for the model.

input_shape
optional shape list, only to be specified if include_top is FALSE (otherwise the input shape has to be (299, 299, 3). It should have exactly 3 inputs channels, and width and height should be no smaller than 75. E.g. (150, 150, 3) would be one valid value.

pooling
Optional pooling mode for feature extraction when include_top is FALSE.

NULL means that the output of the model will be the 4D tensor output of the last convolutional layer.

avg means that global average pooling will be applied to the output of the last convolutional layer, and thus the output of the model will be a 2D tensor.

max means that global max pooling will be applied.

classes
optional number of classes to classify images into, only to be specified if include_top is TRUE, and if no weights argument is specified.

x
Input tensor for preprocessing

Value

A Keras model instance.

Details

On ImageNet, this model gets to a top-1 validation accuracy of 0.790 and a top-5 validation accuracy of 0.945.

Do note that the input image format for this model is different than for the VGG16 and ResNet models (299x299 instead of 224x224).

The xception_preprocess_input() function should be used for image preprocessing.

This application is only available when using the TensorFlow back-end.

Reference

Xception: Deep Learning with Depthwise Separable Convolutions