Install Keras and the TensorFlow backend

Install Keras and the TensorFlow backend

install_keras(method = c("auto", "virtualenv", "conda", "system"),
  conda = "auto")

Arguments

method
Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows (as this isn't supported by TensorFlow). Note also that since this command runs without privillege the "system" method is available only on Windows.

conda
Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).

Note

If you want to do a more customized installation of TensorFlow (including installing a version that takes advantage of Nvidia GPUs if you have the correct CUDA libraries installed) you can call the install_tensorflow() function manually before calling install_keras(). See the article on TensorFlow installation to learn about more advanced options.

See also

install_tensorflow()