layer_dot {keras}	R Documentation

Layer that computes a dot product between samples in two tensors.

Description

Layer that computes a dot product between samples in two tensors.

Usage

layer_dot(inputs, axes, normalize = FALSE, ...)

Arguments

inputs
A list of input tensors (at least 2). Can be missing.

axes
Integer or list of integers, axis or axes along which to take the dot product.

normalize
Whether to L2-normalize samples along the dot product axis before taking the dot product. If set to TRUE, then the output of the dot product is the cosine proximity between the two samples.

...
Standard layer arguments (must be named).

Value

If inputs is supplied: A tensor, the dot product of the samples from the inputs. If inputs is missing, a keras layer instance is returned.

See Also

https://www.tensorflow.org/api_docs/python/tf/keras/layers/dot

https://www.tensorflow.org/api_docs/python/tf/keras/layers/Dot

https://keras.io/api/layers/merging_layers/dot/

Other merge layers: layer_average(), layer_concatenate(), layer_maximum(), layer_minimum(), layer_multiply(), layer_subtract()

[Package keras version 2.6.1 Index]