Turns positive integers (indexes) into dense vectors of fixed size.

For example, list(4L, 20L) -> list(c(0.25, 0.1), c(0.6, -0.2)) This layer can only be used as the first layer in a model.

layer_embedding(object, input_dim, output_dim,
  embeddings_initializer = "uniform", embeddings_regularizer = NULL,
  activity_regularizer = NULL, embeddings_constraint = NULL,
  mask_zero = FALSE, input_length = NULL, batch_size = NULL,
  name = NULL, trainable = NULL, weights = NULL)

Arguments

object
Model or layer object

input_dim
int > 0. Size of the vocabulary, i.e. maximum integer index + 1.

output_dim
int >= 0. Dimension of the dense embedding.

embeddings_initializer
Initializer for the embeddings matrix.

embeddings_regularizer
Regularizer function applied to the embeddings matrix.

activity_regularizer
activity_regularizer

embeddings_constraint
Constraint function applied to the embeddings matrix.

mask_zero
Whether or not the input value 0 is a special "padding" value that should be masked out. This is useful when using recurrent layers, which may take variable length inputs. If this is TRUE then all subsequent layers in the model need to support masking or an exception will be raised. If mask_zero is set to TRUE, as a consequence, index 0 cannot be used in the vocabulary (input_dim should equal size of vocabulary + 1).

input_length
Length of input sequences, when it is constant. This argument is required if you are going to connect Flatten then Dense layers upstream (without it, the shape of the dense outputs cannot be computed).

batch_size
Fixed batch size for layer

name
An optional name string for the layer. Should be unique in a model (do not reuse the same name twice). It will be autogenerated if it isn't provided.

trainable
Whether the layer weights will be updated during training.

weights
Initial weights for layer.

Input shape

2D tensor with shape: (batch_size, sequence_length).

Output shape

3D tensor with shape: (batch_size, sequence_length, output_dim).

References

A Theoretically Grounded Application of Dropout in Recurrent Neural Networks