Wraps arbitrary expression as a layer

Wraps arbitrary expression as a layer

layer_lambda(x, f, mask = NULL, arguments = NULL, input_shape = NULL)

Arguments

x
Model or layer

f
The function to be evaluated. Takes input tensor as first argument.

mask
mask

arguments
optional named list of keyword arguments to be passed to the function.

input_shape
Dimensionality of the input (integer) not including the samples axis. This argument is required when using this layer as the first layer in a model.

Input shape

Arbitrary. Use the keyword argument input_shape (list of integers, does not include the samples axis) when using this layer as the first layer in a model.

Output shape

Arbitrary (based on tensor returned from the function)

See also

Other core layers: layer_activation, layer_activity_regularization, layer_dense, layer_dropout, layer_flatten, layer_input, layer_masking, layer_permute, layer_repeat_vector, layer_reshape