Max pooling operation for spatial data.

Max pooling operation for spatial data.

layer_max_pooling_2d(object, pool_size = c(2L, 2L), strides = NULL,
  padding = "valid", data_format = NULL, batch_size = NULL, name = NULL,
  trainable = NULL, weights = NULL)

Arguments

object
Model or layer object

pool_size
integer or list of 2 integers, factors by which to downscale (vertical, horizontal). (2, 2) will halve the input in both spatial dimension. If only one integer is specified, the same window length will be used for both dimensions.

strides
Integer, list of 2 integers, or NULL. Strides values. If NULL, it will default to pool_size.

padding
One of "valid" or "same" (case-insensitive).

data_format
A string, one of channels_last (default) or channels_first. The ordering of the dimensions in the inputs. channels_last corresponds to inputs with shape (batch, height, width, channels) while channels_first corresponds to inputs with shape (batch, channels, height, width). It defaults to the image_data_format value found in your Keras config file at ~/.keras/keras.json. If you never set it, then it will be "channels_last".

batch_size
Fixed batch size for layer

name
An optional name string for the layer. Should be unique in a model (do not reuse the same name twice). It will be autogenerated if it isn't provided.

trainable
Whether the layer weights will be updated during training.

weights
Initial weights for layer.

Input shape

If data_format='channels_last': 4D tensor with shape: (batch_size, rows, cols, channels)

If data_format='channels_first': 4D tensor with shape: (batch_size, channels, rows, cols)

Output shape

If data_format='channels_last': 4D tensor with shape: (batch_size, pooled_rows, pooled_cols, channels)

If data_format='channels_first': 4D tensor with shape: (batch_size, channels, pooled_rows, pooled_cols)

See also

Other pooling layers: layer_average_pooling_1d, layer_average_pooling_2d, layer_average_pooling_3d, layer_global_average_pooling_1d, layer_global_average_pooling_2d, layer_global_average_pooling_3d, layer_global_max_pooling_1d, layer_global_max_pooling_2d, layer_global_max_pooling_3d, layer_max_pooling_1d, layer_max_pooling_3d