metric_sum {keras}	R Documentation

Computes the (weighted) sum of the given values

Description

Computes the (weighted) sum of the given values

Usage

metric_sum(..., name = NULL, dtype = NULL)

Arguments

...
Passed on to the underlying metric. Used for forwards and backwards compatibility.

name
(Optional) string name of the metric instance.

dtype
(Optional) data type of the metric result.

Details

For example, if values is c(1, 3, 5, 7) then the sum is 16. If the weights were specified as c(1, 1, 0, 0) then the sum would be 4.

This metric creates one variable, total, that is used to compute the sum of values. This is ultimately returned as sum.

If sample_weight is NULL, weights default to 1. Use sample_weight of 0 to mask values.

Value

A (subclassed) Metric instance that can be passed directly to compile(metrics = ), or used as a standalone object. See ?Metric for example usage.

See Also

Other metrics: custom_metric(), metric_accuracy(), metric_auc(), metric_binary_accuracy(), metric_binary_crossentropy(), metric_categorical_accuracy(), metric_categorical_crossentropy(), metric_categorical_hinge(), metric_cosine_similarity(), metric_false_negatives(), metric_false_positives(), metric_hinge(), metric_kullback_leibler_divergence(), metric_logcosh_error(), metric_mean_absolute_error(), metric_mean_absolute_percentage_error(), metric_mean_iou(), metric_mean_relative_error(), metric_mean_squared_error(), metric_mean_squared_logarithmic_error(), metric_mean_tensor(), metric_mean_wrapper(), metric_mean(), metric_poisson(), metric_precision_at_recall(), metric_precision(), metric_recall_at_precision(), metric_recall(), metric_root_mean_squared_error(), metric_sensitivity_at_specificity(), metric_sparse_categorical_accuracy(), metric_sparse_categorical_crossentropy(), metric_sparse_top_k_categorical_accuracy(), metric_specificity_at_sensitivity(), metric_squared_hinge(), metric_top_k_categorical_accuracy(), metric_true_negatives(), metric_true_positives()

[Package keras version 2.6.1 Index]