Provide a scope with mappings of names to custom objects

Provide a scope with mappings of names to custom objects

with_custom_object_scope(objects, expr)

Arguments

objects
Named list of objects

expr
Expression to evaluate

Examples

# NOT RUN {
# define custom metric
sparse_top_k_cat_acc <- function(y_pred, y_true) {
  metric_sparse_top_k_categorical_accuracy(y_pred, y_true, k = 5)
}

with_custom_object_scope(c(top_k_acc = sparse_top_k_cat_acc), {

  # ...define model...

  # compile model (refer to "top_k_acc" by name)
  model %>% compile(
    loss = "binary_crossentropy",
    optimizer = optimizer_nadam(),
    metrics = c("top_k_acc")
  )

  # save the model
  save_model_hdf5("my_model.h5")
  # loading the model within the custom object scope doesn't
  # require explicitly providing the custom_object
  load_model_hdf5("my_model.h5")
})
# }